+++ packages/react-hook-form-schema/src/form/index.ts
@@ -1,2 +1,3 @@
+export { defaultFieldTypes } from "./defaultFieldTypes";
 export { defaultRenderFields, Form, type FormProps } from "./Form";
 export type { DefaultFieldSchema } from "./schema";